requirements.txt
# Financial Orchestrator Dependencies
# Python 3.8+

# Core dependencies
pyyaml>=6.0
jsonschema>=4.0.0
requests>=2.28.0
psutil>=5.9.0

# VN.PY dependencies
vnpy>=4.0.0
vnpy_ctastrategy>=1.4.0
websockets>=10.0.0

# Phase 1: Quant Lab Data Infrastructure
redis>=5.0.0
duckdb>=0.9.0
PyJWT>=2.8.0
cryptography>=41.0.0

# ML Dependencies
scikit-learn>=1.3.0
hmmlearn>=0.3.0

# Web Dashboard
flask>=3.0.0

# Telegram Bot
python-telegram-bot>=20.0

# Async & Live Trading
aiohttp>=3.9.0
python-dotenv>=1.0.0
